deeplab-test-script/requirements.txt
numpy
pillow
matplotlib
thrift
cython
scikit-image
protobuf